Add system property to disable book size limits
If anyone comes in with a watchdog crash related to books after this patch
you will not only be publicly shamed but also made an example of.

Disables the security limits on books entirely, allowing plugins AND players
to make books with as much data as they want. Do not use this without
limiting incoming data from packets in some other way.